pyproject.toml
[project]
name = "plugins-dev-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Umbre11as",email = "olegumbre11a@yandex.ru"}
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "plugins_dev_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
